Move Doctrine type to custom namespace
